+++ src/Document/Service/PageSnippetService.php
@@ -15,28 +15,16 @@ namespace Pimcore\Bundle\StudioBackendBundle\Document\Service;
 
 use Exception;
 use Pimcore\Bundle\StudioBackendBundle\Document\Event\PreResponse\ControllerEvent;
-use Pimcore\Bundle\StudioBackendBundle\Document\Event\PreResponse\RenderBlockEditmodeEvent;
 use Pimcore\Bundle\StudioBackendBundle\Document\Event\PreResponse\TemplateEvent;
-use Pimcore\Bundle\StudioBackendBundle\Document\MappedParameter\RenderAreaBlockParameter;
 use Pimcore\Bundle\StudioBackendBundle\Document\Schema\PageSnippet\Controller;
-use Pimcore\Bundle\StudioBackendBundle\Document\Schema\PageSnippet\RenderAreaBlockData;
 use Pimcore\Bundle\StudioBackendBundle\Document\Schema\PageSnippet\Template;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ElementSavingFailedException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\InvalidArgumentException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ReflectionException;
 use Pimcore\Bundle\StudioBackendBundle\Security\Service\SecurityServiceInterface;
 use Pimcore\Controller\Config\ControllerDataProvider;
-use Pimcore\Document\Editable\Block\BlockStateStack;
-use Pimcore\Document\Editable\EditmodeEditableDefinitionCollector;
-use Pimcore\Http\Request\Resolver\DocumentResolver;
-use Pimcore\Http\Request\Resolver\EditmodeResolver;
-use Pimcore\Localization\LocaleServiceInterface;
-use Pimcore\Model\Document\Editable\Areablock;
 use Pimcore\Model\Document\PageSnippet;
-use Pimcore\Templating\Renderer\EditableRenderer;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\HttpFoundation\Request;
-use Twig\Environment;
 use function sprintf;
 
 /**
@@ -45,15 +33,9 @@ use function sprintf;
 final readonly class PageSnippetService implements PageSnippetServiceInterface
 {
     public function __construct(
-        private BlockStateStack $blockStateStack,
         private ControllerDataProvider $controllerDataProvider,
         private DocumentServiceInterface $documentService,
-        private DocumentResolver $documentResolver,
-        private EditmodeEditableDefinitionCollector $definitionCollector,
-        private EditableRenderer $editableRenderer,
-        private Environment $twig,
         private EventDispatcherInterface $eventDispatcher,
-        private LocaleServiceInterface $localeService,
         private SecurityServiceInterface $securityService,
     ) {
     }
@@ -100,36 +82,6 @@ final readonly class PageSnippetService implements PageSnippetServiceInterface
     /**
      * {@inheritdoc}
      */
-    public function renderAreaBlock(
-        int $documentId,
-        Request $request,
-        RenderAreaBlockParameter $parameter
-    ): RenderAreaBlockData {
-        $document = $this->getPageSnippetClone($documentId);
-        $this->blockStateStack->loadArray($parameter->getBlockStateStack());
-
-        $document->setEditables([]);
-        $this->documentResolver->setDocument($request, $document);
-
-        $this->twig->addGlobal('document', $document);
-        $this->twig->addGlobal('editmode', true);
-
-        $request->attributes->set(EditmodeResolver::ATTRIBUTE_EDITMODE, true);
-        $this->localeService->setLocale($document->getProperty('language'));
-        $html = $this->getHtmlCode($document, $parameter);
-        $data = new RenderAreaBlockData(
-            $this->definitionCollector->getDefinitions(),
-            $html
-        );
-
-        $this->eventDispatcher->dispatch($data, RenderBlockEditmodeEvent::EVENT_NAME);
-
-        return $data;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function setMainDocument(int $documentId, ?string $mainDocumentPath = null): void
     {
         try {
@@ -143,9 +95,9 @@ final readonly class PageSnippetService implements PageSnippetServiceInterface
     }
 
     /**
-     * @throws InvalidArgumentException
+     * {@inheritdoc}
      */
-    private function getPageSnippet(int $id): PageSnippet
+    public function getPageSnippet(int $id): PageSnippet
     {
         $document = $this->documentService->getDocumentElement($this->securityService->getCurrentUser(), $id);
 
@@ -154,61 +106,5 @@ final readonly class PageSnippetService implements PageSnippetServiceInterface
         }
 
         return $document;
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    private function getPageSnippetClone(int $id): PageSnippet
-    {
-        return clone $this->getPageSnippet($id);
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    private function getHtmlCode(PageSnippet $document, RenderAreaBlockParameter $parameter): string
-    {
-        $areaBlockConfig = $parameter->getAreaBlockConfig();
-        $realName = $parameter->getRealName();
-
-        try {
-            $block = $this->editableRenderer->getEditable(
-                $document,
-                'areablock',
-                $realName,
-                $areaBlockConfig,
-                true
-            );
-        } catch (Exception $e) {
-            throw new InvalidArgumentException(
-                sprintf('Cannot get areablock editable for realName (%s): %s', $realName, $e->getMessage()),
-                $e
-            );
-        }
-
-        if (!$block instanceof Areablock) {
-            throw new InvalidArgumentException(
-                sprintf('Editable with realName (%s) is not an areablock', $realName)
-            );
-        }
-
-        $block->setRealName($realName);
-        $block->setEditmode(true);
-        $block->setDataFromEditmode($parameter->getAreaBlockData());
-
-        try {
-            return trim($block->renderIndex($parameter->getIndex(), true));
-        } catch (Exception $e) {
-            throw new InvalidArgumentException(
-                sprintf(
-                    'Cannot render areablock with realName (%s) at index (%d): %s',
-                    $realName,
-                    $parameter->getIndex(),
-                    $e->getMessage()
-                ),
-                $e
-            );
-        }
     }
 }
